+++ Events/BodyScannedEvent.cs
@@ -31,8 +31,8 @@ namespace EddiEvents
             VARIABLES.Add("tidallylocked", "True if the body is tidally locked (only available if DSS equipped)");
             VARIABLES.Add("landable", "True if the body is landable (only available if DSS equipped)");
             VARIABLES.Add("atmosphere", "The atmosphere of the body that has been scanned (only available if DSS equipped)");
-            VARIABLES.Add("atmospherecomposition", "The composition of the atmosphere of the body that has been scanned (only available if DSS equipped)");
-            VARIABLES.Add("solidcomposition", "The composition of the body's solids that has been scanned (only available if DSS equipped)");
+            VARIABLES.Add("atmospherecompositions", "The composition of the atmosphere of the body that has been scanned (array of AtmosphereComposition objects) (only available if DSS equipped)");
+            VARIABLES.Add("solidcompositions", "The composition of the body's solids that has been scanned (array of SolidComposition objects) (only available if DSS equipped)");
             VARIABLES.Add("volcanism", "The volcanism of the body that has been scanned (only available if DSS equipped)");
             VARIABLES.Add("distancefromarrival", "The distance in LS from the main star");
             VARIABLES.Add("orbitalperiod", "The number of days taken for a full orbit of the main star");
@@ -77,9 +77,9 @@ namespace EddiEvents
 
         public AtmosphereClass atmosphereclass { get; private set; }
 
-        public List<AtmosphereComposition> atmospherecomposition { get; private set; }
+        public List<AtmosphereComposition> atmospherecompositions { get; private set; }
 
-        public List<SolidComposition> solidcomposition { get; private set; }
+        public List<SolidComposition> solidcompositions { get; private set; }
 
         public Volcanism volcanism { get; private set; }
 
@@ -139,8 +139,8 @@ namespace EddiEvents
             this.tidallylocked = tidallylocked;
             this.landable = landable;
             this.atmosphereclass = atmosphereClass;
-            this.atmospherecomposition = atmosphereComposition;
-            this.solidcomposition = solidCompositions;
+            this.atmospherecompositions = atmosphereComposition;
+            this.solidcompositions = solidCompositions;
             this.volcanism = volcanism;
             this.orbitalperiod = orbitalperiodDays;
             this.rotationperiod = rotationperiodDays;