+++ EDDI/ChangeLog.md
@@ -119,6 +119,9 @@ Full details of the variables available for each noted event, and VoiceAttack in
     * Fixed a bug that could prevent proper lookup of bodies using the `BodyDetails()` function.
     * Added new `Next jump` event
     * Added new top level object `nextsystem` - like `lastsystem` but for the next system you are visiting. e.g. "Our next waypoint is \{nextsystem.name\}".
+    * Revised `Body scanned` event variables for better interchangeability with with the `BodyDetails` function
+      * WAS: `atmospherecomposition`, IS: `atmospherecompositions`
+      * WAS: `solidcomposition`, IS: `solidcompositions`
   * VoiceAttack  
     * Added new `Next system` variables